+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -202,7 +202,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KC5WJPRF7EJGAA46742S7Q9P"),
+  _id: z.string().default("01KC6JYPE007XNYJS5C9QGNGDT"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -245,7 +245,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KC5WJPRF7EJGAA46742S7Q9P"),
+  id: z.string().default("01KC6JYPE007XNYJS5C9QGNGDT"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,